Relocate ROOT step/manager files
