+++ app/demo-loop/demo-loop.cc
@@ -33,14 +33,14 @@
 #include "corecel/sys/ScopedMpiInit.hh"
 #include "corecel/sys/Stopwatch.hh"
 #include "celeritas/Types.hh"
+#include "celeritas/ext/RootFileManager.hh"
 #include "celeritas/ext/ScopedRootErrorHandler.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/io/EventReader.hh"
-#include "celeritas/io/RootFileManager.hh"
-#include "celeritas/io/RootStepWriter.hh"
 #include "celeritas/phys/Primary.hh"
 #include "celeritas/phys/PrimaryGenerator.hh"
 #include "celeritas/phys/PrimaryGeneratorOptions.hh"
+#include "celeritas/user/RootStepWriter.hh"
 #include "celeritas/user/StepCollector.hh"
 #include "celeritas/user/StepData.hh"
