+++ src/celeritas/CMakeLists.txt
@@ -159,9 +159,9 @@ endif()
 if(CELERITAS_USE_JSON)
   list(APPEND SOURCES
     ext/GeantPhysicsOptionsIO.json.cc
-    io/RootStepWriterIO.json.cc
     field/FieldDriverOptionsIO.json.cc
     phys/PrimaryGeneratorOptionsIO.json.cc
+    user/RootStepWriterIO.json.cc
   )
   list(APPEND PRIVATE_DEPS nlohmann_json::nlohmann_json)
 endif()
@@ -204,8 +204,8 @@ if(CELERITAS_USE_ROOT)
     ext/RootImporter.cc
     ext/ScopedRootErrorHandler.cc
     ext/RootUniquePtr.root.cc
-    io/RootFileManager.cc
-    io/RootStepWriter.cc
+    ext/RootFileManager.cc
+    user/RootStepWriter.cc
     "${CMAKE_CURRENT_BINARY_DIR}/CeleritasRootInterface.cxx"
   )
